Adds CLion support
This PR adds a few configuration files for CLion. These configuration files only rename the makefiles targets to be clearer, and specifies a target name to attach the debugger to for each test target.

It also adds the debug symbols (.dSYM) files generated on macOS by llvm to the .gitignore file.