Force usage of LLDB inside of VScode (MIMode)
This commit forces the usage of lldb, as specificed by the VScode
documentation [0]. It should use lldb to begin with, as required by the
course manual [1], but it seems to try to use gdb instead.

Note that gdb does not support aarch64, and thus causes an error on
Apple Silicon devices. This might also be an issue for Linux users.

[0]: https://code.visualstudio.com/docs/cpp/launch-json-reference#_mimode
[1]: https://vu-oofp.gitlab.io/website/manual/framework.html